requirements.txt
python-telegram-bot>=21.0
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
pyngrok>=6.0.0
edge-tts>=7.0.0
